refactor: move generateSampleStoryData function to story file
